backend/requirements.txt
# Build tools (must be first)
setuptools>=65.5.0
wheel>=0.38.4
Cython>=0.29.36

# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
starlette==0.27.0
anyio>=3.7.1,<4.0.0  # Compatible with FastAPI 0.104.1

# Data processing and ML
numpy>=1.26.2
pandas>=2.1.3
scikit-learn>=1.3.2
tensorflow>=2.15.0
yfinance>=0.2.33

# Web and API
beautifulsoup4>=4.12.2
aiohttp>=3.9.1
requests>=2.31.0
python-multipart>=0.0.6

# NLP
textblob>=0.17.1
nltk>=3.8.1

# Visualization
plotly>=5.18.0
ta>=0.10.2

# Auth and DB
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
sqlalchemy>=1.4.49
alembic>=1.12.1
psycopg2-binary>=2.9.9

# Testing and development
pytest>=7.4.3
httpx>=0.25.2
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1
joblib>=1.3.2
finnhub-python>=2.4.18
psutil>=5.9.7
typing-extensions>=4.8.0
idna>=3.6
sniffio>=1.3.0
h11>=0.14.0
httptools>=0.6.1
websockets>=12.0
click>=8.1.7
colorama>=0.4.6
distro>=1.9.0
watchfiles>=0.21.0
